+++ src/SummaryCreator/Configuration/Json/JsonConfigurationConverter.cs
@@ -91,7 +91,7 @@ namespace SummaryCreator.Configuration.Json
 
         public string Sheet { get; set; }
 
-        public long Row { get; set; }
+        public int Row { get; set; }
     }
 
     public class JsonTimeseriesModel
